requirements.txt
filelock==3.20.0
fsspec==2025.12.0
Jinja2==3.1.6
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.6.1
numpy==2.3.5
pillow==12.0.0
python-rtmidi==1.5.8
setuptools==70.2.0
sympy==1.14.0
torch==2.10.0+cpu
torchvision==0.25.0+cpu
typing_extensions==4.15.0
